Refactor version argument
