+++ paifulogger/__init__.py
@@ -8,11 +8,12 @@ del sys
 main_path = os.path.dirname(os.path.abspath(__file__))
 del os
 
+from .version import __version__
+
 from .log import config_path, log, log_paifu, log_parser, remove_old_paifu
 from .paifu_dl import paifu_dl
 from .src.get_paifu import get_paifu
 from .src.i18n import localized_str
-from .version import __version__
 
 __all__ = [
     "config_path",
